Move istio-jwt manifests in separate folder and introduce testsuite files
