+++ tests/integration/jwt_handler.go
@@ -1,0 +1,62 @@
+package api_gateway
+
+import (
+	"fmt"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"strings"
+
+	"github.com/kyma-project/api-gateway/tests/integration/pkg/client"
+)
+
+func SwitchJwtHandler(jwtHandler string) (string, error) {
+	mapper, err := client.GetDiscoveryMapper()
+	if err != nil {
+		return "", err
+	}
+	mapping, err := mapper.RESTMapping(schema.ParseGroupKind("ConfigMap"))
+	if err != nil {
+		return "", err
+	}
+	currentJwtHandler, configMap, err := getConfigMapJwtHandler(mapping.Resource)
+	if err != nil {
+		configMap := unstructured.Unstructured{
+			Object: map[string]interface{}{
+				"kind":       "ConfigMap",
+				"apiVersion": "v1",
+				"metadata": map[string]interface{}{
+					"name":      configMapName,
+					"namespace": defaultNS,
+				},
+				"data": map[string]interface{}{
+					"api-gateway-config": "jwtHandler: " + jwtHandler,
+				},
+			},
+		}
+		currentJwtHandler = jwtHandler
+		err = resourceManager.CreateResource(k8sClient, mapping.Resource, defaultNS, configMap)
+	}
+	if err != nil {
+		return "", fmt.Errorf("could not get or create jwtHandler config:\n %+v", err)
+	}
+	if currentJwtHandler != jwtHandler {
+		configMap.Object["data"].(map[string]interface{})["api-gateway-config"] = "jwtHandler: " + jwtHandler
+		err = resourceManager.UpdateResource(k8sClient, mapping.Resource, defaultNS, configMapName, *configMap)
+		if err != nil {
+			return "", fmt.Errorf("unable to update ConfigMap:\n %+v", err)
+		}
+	}
+	return currentJwtHandler, err
+}
+
+func getConfigMapJwtHandler(gvr schema.GroupVersionResource) (string, *unstructured.Unstructured, error) {
+	res, err := resourceManager.GetResource(k8sClient, gvr, defaultNS, configMapName)
+	if err != nil {
+		return "", res, fmt.Errorf("could not get ConfigMap:\n %+v", err)
+	}
+	data, found, err := unstructured.NestedMap(res.Object, "data")
+	if err != nil || !found {
+		return "", res, fmt.Errorf("could not find data in the ConfigMap:\n %+v", err)
+	}
+	return strings.Split(data["api-gateway-config"].(string), ": ")[1], res, nil
+}
