+++ tests/integration/manifests/istio-jwt/istio-jwt-and-oauth.yaml
