+++ tests/integration/manifests/istio-jwt/istio-jwt-and-unrestricted.yaml
