+++ tests/integration/manifests/istio-jwt/istio-jwt-audiences.yaml
