+++ tests/integration/manifests/istio-jwt/istio-jwt-common.yaml
