+++ tests/integration/manifests/istio-jwt/istio-jwt-custom-label-selector.yaml
