+++ tests/integration/manifests/istio-jwt/istio-jwt-diff-svc-same-methods.yaml
