+++ tests/integration/manifests/istio-jwt/istio-jwt-multiple-mutators.yaml
