+++ tests/integration/manifests/istio-jwt/istio-jwt-mutator-cookie.yaml
