+++ tests/integration/manifests/istio-jwt/istio-jwt-mutator-header.yaml
