+++ tests/integration/manifests/istio-jwt/istio-jwt-mutators-overwrite.yaml
