+++ tests/integration/manifests/istio-jwt/istio-jwt-scopes.yaml
