+++ tests/integration/manifests/istio-jwt/istio-jwt-service-fallback.yaml
