+++ tests/integration/manifests/istio-jwt/istio-jwt-token-from-headers.yaml
