+++ tests/integration/manifests/istio-jwt/istio-jwt-token-from-params.yaml
