+++ tests/integration/manifests/istio-jwt/istio-jwt-two-namespaces.yaml
