+++ tests/integration/manifests/istio-jwt/second-namespace.yaml
