+++ tests/integration/manifests/istio-jwt/testing-app.yaml
