+++ tests/integration/manifests/istio-jwt/testing-helloworld-app.yaml
