+++ tests/integration/manifests/istio-jwt/testing-helloworld-custom-label-app.yaml
