+++ tests/integration/scenario_jwt_and_unrestricted_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func initJwtAndAllow(ctx *godog.ScenarioContext) {
-	s, err := CreateScenarioWithRawAPIResource("istio-jwt-and-unrestricted.yaml", "istio-jwt-unrestricted")
+	s, err := CreateIstioJwtScenario("istio-jwt-and-unrestricted.yaml", "istio-jwt-unrestricted")
 	if err != nil {
 		t.Fatalf("could not initialize scenario err=%s", err)
 	}
