+++ tests/integration/scenario_jwt_token_from_test.go
@@ -7,7 +7,7 @@ import (
 func initTokenFromHeaders(ctx *godog.ScenarioContext) {
 	jwtHeaderName = "x-jwt-token"
 
-	s, err := CreateScenarioWithRawAPIResource("istio-jwt-token-from-headers.yaml", "istio-jwt-token-from-headers")
+	s, err := CreateIstioJwtScenario("istio-jwt-token-from-headers.yaml", "istio-jwt-token-from-headers")
 	if err != nil {
 		t.Fatalf("could not initialize scenario err=%s", err)
 	}
@@ -25,7 +25,7 @@ func initTokenFromHeaders(ctx *godog.ScenarioContext) {
 func initTokenFromParams(ctx *godog.ScenarioContext) {
 	fromParamName = "jwt_token"
 
-	s, err := CreateScenarioWithRawAPIResource("istio-jwt-token-from-params.yaml", "istio-jwt-token-from-params")
+	s, err := CreateIstioJwtScenario("istio-jwt-token-from-params.yaml", "istio-jwt-token-from-params")
 	if err != nil {
 		t.Fatalf("could not initialize scenario err=%s", err)
 	}
