+++ tests/integration/scenario_jwt_two_namespaces_test.go
@@ -5,7 +5,7 @@ import (
 )
 
 func initJwtTwoNamespaces(ctx *godog.ScenarioContext) {
-	s, err := CreateScenarioWithRawAPIResource("istio-jwt-two-namespaces.yaml", "istio-jwt-two-namespaces")
+	s, err := CreateIstioJwtScenario("istio-jwt-two-namespaces.yaml", "istio-jwt-two-namespaces")
 	if err != nil {
 		t.Fatalf("could not initialize scenario err=%s", err)
 	}
