+++ tests/integration/scenario_service_fallback_test.go
@@ -5,7 +5,7 @@ import (
 )
 
 func initJwtServiceFallback(ctx *godog.ScenarioContext) {
-	s, err := CreateScenarioWithRawAPIResource("istio-jwt-service-fallback.yaml", "istio-jwt-service-fallback")
+	s, err := CreateIstioJwtScenario("istio-jwt-service-fallback.yaml", "istio-jwt-service-fallback")
 	if err != nil {
 		t.Fatalf("could not initialize scenario err=%s", err)
 	}
