+++ tests/integration/testsuite_custom_domain_test.go
@@ -1,0 +1,7 @@
+package api_gateway
+
+import "github.com/cucumber/godog"
+
+func initCustomDomainTestsuite(ctx *godog.ScenarioContext) {
+	InitializeScenarioCustomDomain(ctx)
+}
